requirement.txt
playwright
geoip2
pyap2
pandas
beautifulsoup4
flask
